proto: move RRSIG test into rrsig module
